// src/amazhop/Products.enum.ts
export const enum Products {
    AVOCADO_CASE = 15,
    BATTERIES = 30,
    GLASSES = 15,
    T_SHIRT = 20,
    COFFEE_MACHINE = 50,
    PEN_DRIVE = 45,
    PET_BED = 35,
    CHAIR = 85,
    BOXES = 10,
    PILLOW = 29,
    NOTEBOOK = 3,
    PAN = 39,
    SECURITY_BOX = 87,
    DESK = 459,
    BALL = 18,
    BACKPACK = 158
}